Base Series System for Simulation Study

The "base" system in our simulation study in the paper "Reliability Estimation in Series Systems".

alex_weibull_series

Format

A list with the following components:

theta

The shape and scale parameters of the components combined

shapes

The shape parameters of the components

scales

The scale parameters of the components

family

The family of the likelihood model (weibull series)

Details

It has five components in series. It is based on the example in Guo, table 2, but with two new components added. Each component has approximately the same reliability, so that there is no clear weakest link. The series system parameter vector is given by:

theta = (shape1 = 1.2576, scale1 = 994.3661,
         shape2 = 1.1635, scale2 = 908.9458,
         shape3 = 1.1308, scale3 = 840.1141,
         shape4 = 1.1802, scale4 = 940.1342,
         shape5 = 1.2034, scale5 = 923.1631)

See also