Conditional candidate set sampler when component cause of failure is known and the system lifetime is observed. (UNTESTED)

Conditional sampler Ci | Ti = t, Ki = k We have a smoothing parameter that specifies the bin's width for the discretization of the system lifetimes.

conditional_masked_cause(n, df, t, k, nbins = 10)

Arguments

df

data frame (sample) that we used to estimate Ci | Ti = ti, Ki = j df should only contain data in which the system failed, rather than being right-censored.

t

observed lifetime, defaults to NA (unknown or unconditional)

k

component cause of failure, defaults to NA (unknown)

nbins

number of bins to use for discretizing the component lifetimes

Details

If component cause of failure is known, then we can sample from the conditional distribution of Ci | Ti = t, Ki = k, otherwise we can sample from the empirical distribution of Ci | Ti = ti, which may still be reasonable, since by Condition 2, PrCi = ci | Ti = ti, Ki = j = PrCi = ci | Ti = ti, Ki = j' for all j, j' in ci. Of course, we may violate Condition 1, PrKi in Ci = 1.

Note that if |ci| = 1, then we can sample from the empirical distribution of Ci | Ti = ti, Ki = j, since there is only one possible value for Ki, Ki = j if ci = j. Otherwise, if |ci| > 1, and we don't know the component cause of failure, then we can sample from the empirical distribution of Ci | Ti = ti.

In the semi-parametric bootstrap, we generate the samples ourselves from our estimate of theta, so we can sample from the conditional distribution of Ci | Ti = ti, Ki = j, since we know the simulated component cause of failure. This is the main reason we want to sample Ci | Ti = ti, Ki = j, so this is fine.