Example Data for a Series System (Table 2, Guo 2013)

The source of this data set comes from "Estimating Component Reliabilities from Incomplete System Failure Data", Table 2: Example Data for a Series System.

guo_weibull_series_table_2

Format

A data frame with 30 rows and 4 columns:

t

number, lifetime of series system

x1

logical, TRUE indicates component 1 is in the candidate set, FALSE otherwise

x2

logical, TRUE indicates component 2 is in the candidate set, FALSE otherwise

x3

logical, TRUE indicates component 3 is in the candidate set, FALSE otherwise

delta

logical, Right-censoring indicator, TRUE if the system is observed

Source

H. Guo, F. Szidarovszky, and P. Niu, "Estimating component reliabilities from incomplete system failure data," in 2013 Proceedings Annual Reliability and Maintainability Symposium (RAMS), 2013. Online. Available: https://doi.org/10.1109/rams.2013.6517765

Details

When you use likelihood model that assumes a Weibull series system and candidate sets represented by Boolean vectors (x1, x2, x3) that satisfy conditions C1, C2, and C3, the MLE of the shape and scale parameters are:

β1 = 1.2576

η1 = 994.3661

β2 = 1.1635

η2 = 908.9458

β3 = 1.1308

η3 = 840.1141

θ̂ = (β1, η1, β2, η2, β3, η3)

This has a log-likelihood of -228.6851.

See also

Examples

head(guo_weibull_series_md)
#> $mle
#> [1]   1.2576 994.3661   1.1635 908.9458   1.1308 840.1141
#> 
#> $shape_mles
#> [1] 1.2576 1.1635 1.1308
#> 
#> $scale_mles
#> [1] 994.3661 908.9458 840.1141
#> 
#> $loglike
#> [1] -228.6851
#> 
#> $p_hat
#> [1] 0.215
#> 
#> $tau
#> [1] Inf
#> 
ll <- loglik_wei_series_md_c1_c2_c3(df = guo_weibull_series_table_2)
#> Error in loglik_wei_series_md_c1_c2_c3(df = guo_weibull_series_table_2): argument "theta" is missing, with no default
sol <- optim(par = guo_weibull_series_md$mle,
             fn = ll,
             hessian = TRUE,
             control = list(parscale = c(1, 1000, 1, 1000, 1, 1000)))
#> Error in (function (par) fn(par, ...))(c(1.2576, 994.3661, 1.1635, 908.9458, 1.1308, 840.1141)): object 'll' not found
abs(sol$value - guo_weibull_series_md$loglike) < 1e-4
#> Error in eval(expr, envir, enclos): object 'sol' not found
abs(sol$par - guo_weibull_series_md$mle) < 1e-4
#> Error in eval(expr, envir, enclos): object 'sol' not found