Generates a log-likelihood function for a Weibull series system with respect to parameter theta (shape, scale) for masked data with candidate sets that satisfy conditions C1, C2, and C3 and right-censored data.

Generates a log-likelihood function for a Weibull series system with respect to parameter theta (shape, scale) for masked data with candidate sets that satisfy conditions C1, C2, and C3 and right-censored data.

loglik_wei_series_md_c1_c2_c3(
  df,
  theta,
  candset = "x",
  lifetime = "t",
  right_censoring_indicator = "delta"
)

Arguments

df

(masked) data frame

theta

parameter vector (shape1, scale1, ..., shapem, scalem)

candset

prefix of Boolean matrix encoding of candidate sets, defaults to x, e.g., x1,...,xm.

lifetime

system lifetime (optionally right-censored) column name

right_censoring_indicator

right-censoring indicator column name, if TRUE, then the system lifetime is observed, otherwise it is right-censored. If there is no right-censoring indicator column by the given name, then the system lifetimes are assumed to be observed.

Value

Log-likelihood with respect to theta given df