Nelder-Mead MLE solver for the Weibull series model with masked component cause of failure with observation of exact system lifetime and right censoring.

Nelder-Mead MLE solver for the Weibull series model with masked component cause of failure with observation of exact system lifetime and right censoring.

mle_nelder_wei_series_md_c1_c2_c3(
  df,
  theta0,
  hessian = TRUE,
  ...,
  control = list()
)

Arguments

df

data frame, right-censored lifetimes with masked component cause of failure

theta0

initial parameter vector

hessian

logical, if TRUE, then the Hessian is estimated and returned

...

additional arguments passed to loglik_wei_series_md_c1_c2_c3.

control

list of control parameters passed to optim for `Nelder-Mead'

Value

list with components:

par final parameter vector

value final log-likelihood

counts number of function evaluations, gradient evaluations, and Hessian evaluations

convergence convergence code

message convergence message

hessian estimated Hessian